+++ {{ cookiecutter.project_name }}/django-apps/project_utils/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from .views import ExposedProjectGlobalsView
+from .views import ExposedProjectGlobalsView, StackInfoView
 
 
 app_name = "project_utils"
@@ -11,5 +11,10 @@ urlpatterns = [
         "project-globals/",
         ExposedProjectGlobalsView.as_view(),
         name="project-globals"
+    ),
+    path(
+        "stack-info/",
+        StackInfoView.as_view(),
+        name="stack-info"
     ),
 ]