+++ {{ cookiecutter.project_name }}/django-apps/project_utils/views/__init__.py
@@ -1,6 +1,8 @@
 from .project_globals import ExposedProjectGlobalsView
+from .stack_info import StackInfoView
 
 
 __all__ = [
     "ExposedProjectGlobalsView",
+    "StackInfoView",
 ]